[mesh] refactoring
